refactor : 테스트 패키지 구조도 메인과 동일하게 변경
