--- src/test/java/lotto/domain/lottos/LottoBoughtTicketTest.java
@@ -1,4 +1,0 @@
-import static org.junit.jupiter.api.Assertions.*;
-class LottoBoughtTicketTest {
-  
-}
