+++ src/test/java/lotto/domain/lottos/LottoNumberTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lottos;
 
 import lotto.domain.lottos.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
