+++ src/test/java/lotto/domain/lottos/winnerlotto/LottoWinnerBonusNumberTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lottos.winnerlotto;
 
 import lotto.domain.lottos.winnerlotto.LottoWinnerTicket;
 import lotto.domain.lottos.LottoNumber;
